Added discount visual to cards and updated file name from productListing to ProductList
