+++ src/js/main.js
@@ -1,5 +1,5 @@
 import ProductData from "./ProductData.mjs";
-import ProductListing from "./ProductListing.mjs";
+import ProductListing from "./ProductList.mjs";
 
 const data = new ProductData("tents");
 const element = document.querySelector(".product-list");
